+++ lib/game_interface/text_interface_component.dart
@@ -37,4 +37,7 @@ class TextInterfaceComponent extends InterfaceComponent {
       Vector2(x, y),
     );
   }
+
+  @override
+  bool get isVisible => true;
 }